pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
    "setuptools_scm[toml]>=7"
]
build-backend = "setuptools.build_meta"

[project]
name = "reboost"
description = "New LEGEND Monte-Carlo simulation post-processing"
authors = [
    { name = "Manuel Huber", email = "info@manuelhu.de" },
    { name = "Toby Dixon", email = "toby.dixon.23@ucl.ac.uk" },
    { name = "Luigi Pertoldi", email = "gipert@pm.me" },
]
maintainers = [
    { name = "The LEGEND Collaboration" },
]
readme = "README.md"
license.file = "LICENSE"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: MacOS",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.9"
dependencies = [
    "colorlog",
    "numpy",
    "scipy",
    "numba",
    "legend-pydataobj>=1.11.6",
    "legend-pygeom-optics>=0.6.5",
    "hist",
    "dbetto",
    "particle",
    "pandas",
    "matplotlib",
    "pygama",
    "pyg4ometry",
]
dynamic = [
    "version",
]

[project.urls]
Homepage = "https://github.com/legend-exp/reboost"
"Bug Tracker" = "https://github.com/legend-exp/reboost/issues"
Discussions = "https://github.com/legend-exp/reboost/discussions"
Changelog = "https://github.com/legend-exp/reboost/releases"

[tool.setuptools_scm]
write_to = "src/reboost/_version.py"

[project.optional-dependencies]
all = [
    "reboost[docs,test]",
]
docs = [
    "furo",
    "myst-parser",
    "sphinx",
    "sphinx-copybutton",
]
test = [
    "pre-commit",
    "pytest>=6.0",
    "pytest-cov",
    "legend-pygeom-hpges",
    "legend-pygeom-tools",
    "pyg4ometry",
    "pylegendtestdata>=0.6",

]

[project.scripts]
reboost-optical = "reboost.optmap.cli:optical_cli"
reboost         = "reboost.cli:cli"

[tool.setuptools]
include-package-data = true
zip-safe = false
license-files = [
    "LICENSE",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [
    "src",
]
namespaces = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = "error"
log_cli_level = "info"
testpaths = "tests"

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "ARG",      # flake8-unused-arguments
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
  "D",        # Docstrings
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
  "PT011",    # too picky pytest.raises() complaint
  "D1",       # It's ok to have undocumented things
  "D203",     # blank line required before class docstring
  "D213",     # Multi-line docstring summary should start at the first line
  "D401",     # Summary does not need to be in imperative mood
  "D413",     # No blank line after last section in docstring
  "PLC2401",  # We like non-ASCII characters for math
]
isort.required-imports = ["from __future__ import annotations"]
# see also napoleon config in docs/source/conf.py
pydocstyle.convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]

[tool.docformatter]
black = true
wrap-summaries = 9999
